requirements.txt
-i https://pypi.org/simple
cachetools==4.1.1; python_version ~= '3.5'
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
demjson==2.2.4
flask==1.1.2
google-api-core==1.21.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-auth==1.19.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-cloud-core==1.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-cloud-storage==1.29.0
google-resumable-media==0.5.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
googleapis-common-protos==1.52.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
itsdangerous==1.1.0
jinja2==2.11.2
markupsafe==1.1.1
mmfont==0.1
myanmartools==0.1
numpy==1.19.0; python_version >= '3.6'
pillow==7.2.0
protobuf==3.12.2
pyasn1-modules==0.2.8
pyasn1==0.4.8
pytz==2020.1
requests==2.24.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rsa==4.6; python_version >= '3'
shortuuid==1.0.1
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.25.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
werkzeug==1.0.1
